After thinking and trying for a long time, it is decided that we won't check if node's content is dirty: it can not be done in immediate mode. See Box.cs, node's content will be refreshed every frame. Implemented comparison of dependentResources (Pen, Brush, Geometry, GlyphRun and FormattedText) and primitive types (Point and Rect). Implemented refreshing content: Only update binary records of a content that has been changed. Refactored: extracted ContentCheckerBase. Arrange names of record types' field. Added content reading methods to read one single record.
bugs Introduced:
* builtin text rendering is broken; pixel perfect rendering is broken, for example BuiltinPrimitiveRenderer.DrawLine
* template program won't run because controls hasn't been rewritten as to refresh its content every frame.
