+++ src/ImGui/OSAbstraction/Text/FormattedText.cs
@@ -31,12 +31,37 @@ namespace ImGui.OSAbstraction.Text
 
         public FormattedText(Point origin, string text, string fontFamily, double fontSize)
         {
+            //pre-compute hash code
+            unchecked
+            {
+                int hash = 17;
+                hash = hash * 23 + text.GetHashCode();
+                hash = hash * 23 + fontFamily.GetHashCode();
+                hash = hash * 23 + fontSize.GetHashCode();
+                this.hashCode = hash;
+            }
+
             Initialize(origin, text, fontFamily, fontSize);
+        }
+
+        public override bool Equals(object obj)
+        {
+            GlyphRun other = obj as GlyphRun;
+            if (other == null)
+            {
+                return false;
+            }
+
+            return other.GetHashCode() == GetHashCode();
+        }
+
+        public override int GetHashCode()
+        {
+            return this.hashCode;
         }
 
         private void Initialize(Point origin, string s, string fontFamily, double fontSize)
         {
-
             var textContext = new OSImplentation.TypographyTextContext(s, fontFamily, fontSize, TextAlignment.Leading);
             textContext.Build(origin);
 
@@ -60,5 +85,6 @@ namespace ImGui.OSAbstraction.Text
         }
 
         private string text;
+        private readonly int hashCode;
     }
 }
