+++ test/ImGui.UnitTest/Util/Util.cs
@@ -4,17 +4,13 @@ using System.Diagnostics;
 using System.IO;
 using System.Reflection;
 using System.Runtime.CompilerServices;
-using Cairo;
 using ImageSharp.Extension;
-using ImGui.GraphicsAbstraction;
 using ImGui.GraphicsImplementation;
 using ImGui.Rendering;
 using SixLabors.ImageSharp;
 using SixLabors.ImageSharp.PixelFormats;
 using SixLabors.ImageSharp.Processing;
 using Xunit;
-using Image = SixLabors.ImageSharp.Image;
-using Path = System.IO.Path;
 
 namespace ImGui.UnitTest
 {
@@ -93,8 +89,8 @@ namespace ImGui.UnitTest
 
         internal static void DrawNode(Node node, [CallerMemberName] string memberName = "")
         {
-            using (ImageSurface surface = new ImageSurface(Format.Argb32, (int)node.Rect.Width, (int)node.Rect.Height))
-            using (Context context = new Context(surface))
+            using (Cairo.ImageSurface surface = new Cairo.ImageSurface(Cairo.Format.Argb32, (int)node.Rect.Width, (int)node.Rect.Height))
+            using (Cairo.Context context = new Cairo.Context(surface))
             {
                 Draw(context, node);
 
@@ -109,7 +105,7 @@ namespace ImGui.UnitTest
             }
         }
 
-        private static void Draw(Context context, Visual visual)
+        private static void Draw(Cairo.Context context, Visual visual)
         {
             var node = (Node)visual;
             var isGroup =  node.IsGroup;
@@ -212,12 +208,11 @@ namespace ImGui.UnitTest
 
             MeshBuffer meshBuffer = new MeshBuffer();
             MeshList meshList = new MeshList();
-            IGeometryRenderer geometryRenderer = new BuiltinGeometryRenderer();
-
+            BuiltinGeometryRenderer geometryRenderer = new BuiltinGeometryRenderer();
             using (var context = new RenderContextForTest(width, height))
             {
                 //This must be called after the context is created, for uploading textures to GPU via OpenGL.
-                node.Draw(geometryRenderer, meshList);
+                node.Render(new RenderContext(geometryRenderer, meshList));
 
                 //rebuild mesh buffer
                 meshBuffer.Clear();
@@ -236,14 +231,14 @@ namespace ImGui.UnitTest
         {
             MeshBuffer meshBuffer = new MeshBuffer();
             MeshList meshList = new MeshList();
-            IGeometryRenderer geometryRenderer = new BuiltinGeometryRenderer();
+            BuiltinGeometryRenderer geometryRenderer = new BuiltinGeometryRenderer();
 
             using (var context = new RenderContextForTest(width, height))
             {
                 //This must be called after the context is created, for uploading textures to GPU via OpenGL.
                 foreach (var node in nodes)
                 {
-                    node.Draw(geometryRenderer, meshList);
+                    node.Render(new RenderContext(geometryRenderer, meshList));
                 }
 
                 //rebuild mesh buffer
@@ -263,12 +258,12 @@ namespace ImGui.UnitTest
         {
             MeshBuffer meshBuffer = new MeshBuffer();
             MeshList meshList = new MeshList();
-            var primitiveRenderer = new BuiltinGeometryRenderer();
+            BuiltinGeometryRenderer geometryRenderer = new BuiltinGeometryRenderer();
 
             using (var context = new RenderContextForTest(width, height))
             {
                 //This must be called after the context is created, for uploading textures to GPU via OpenGL.
-                root.Foreach(n =>primitiveRenderer.Draw(n, clipRect, meshList));
+                root.Render(new RenderContext(geometryRenderer, meshList));
 
                 //rebuild mesh buffer
                 meshBuffer.Clear();
